batch_solution/requirements.txt
pandas
mlflow
python-dotenv
loguru
boto3
